server/observability/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
opentelemetry-distro==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-sdk==1.21.0
psutil==5.9.6